exception handling für keyword und publicationtype

+++ IAA/src/de/nak/librarymgmt/service/KeywordServiceImpl.java
@@ -1,47 +1,45 @@
 package de.nak.librarymgmt.service;
 
-import java.util.HashSet;
 import java.util.Set;
 
 import de.nak.librarymgmt.dao.KeywordDAO;
 import de.nak.librarymgmt.model.Keyword;
 
+import org.springframework.dao.DataIntegrityViolationException;
+
 public class KeywordServiceImpl implements KeywordService {
 
 	private KeywordDAO keywordDAO;
 
-
+	/** {@inheritDoc} */
 	@Override
-	public void createKeyword(String name) {
+	public void createKeyword(String name) throws Exception {
 		Keyword keyword = new Keyword();
 		keyword.setName(name);
 		try {
 			keywordDAO.save(keyword);
-		} catch (Exception e) {
-			// TODO
+		} catch (DataIntegrityViolationException ex) {
 		}
-
 	}
 
 	@Override
-	public void deleteKeyword(String name) {
+	public void deleteKeyword(String name) throws NotFoundException {
+		keywordDAO.getSessionFactory().getCurrentSession().beginTransaction();
 		Keyword keyword = keywordDAO.findByName(name);
-		try {
-			keywordDAO.delete(keyword);
-		} catch (Exception e) {
-			// TODO
+		if (keyword == null) {
+			throw new NotFoundException("Keyword does not exist");
 		}
-
+		keywordDAO.delete(keyword);
 	}
 
 	@Override
-	public void updateKeyword(String name, String newName) {
+	public void updateKeyword(String name, String newName)
+			throws NotFoundException {
 		Keyword keyword = keywordDAO.findByName(name);
-		try {
-			keyword.setName(newName);
-		} catch (Exception e) {
-			// TODO
+		if (keyword == null) {
+			throw new NotFoundException("Keyword does not exist");
 		}
+		keyword.setName(newName);
 
 	}
 
@@ -51,11 +49,13 @@ public class KeywordServiceImpl implements KeywordService {
 	}
 
 	@Override
-	public Keyword findKeywordByName(String name) {
+	public Keyword findKeywordByName(String name) throws NotFoundException {
 		Keyword keyword = keywordDAO.findByName(name);
+		if (keyword == null) {
+			throw new NotFoundException("Keyword does not exist");
+		}
 		return keyword;
 	}
-
 
 	public void setKeywordDAO(KeywordDAO keywordDAO) {
 		this.keywordDAO = keywordDAO;
